separate native and bus linker, implement single receive for bus lookup
